things about back in order

+++ pkgs/servers/http/conveyor-apache/builder.sh
@@ -15,28 +15,18 @@ cd httpd-*
 make
 make install
 
+mkdir -p /home/user/.conveyor/runtime/dogfoodsoftware.com/conveyor/distro/pkgs/servers/http/conveyor-apache
+rm /home/user/.conveyor/runtime/dogfoodsoftware.com/conveyor/distro/pkgs/servers/http/conveyor-apache/runnable
+ln -s $out/ /home/user/.conveyor/runtime/dogfoodsoftware.com/conveyor/distro/pkgs/servers/http/conveyor-apache/runnable
+
 # Set up the static Conveyor-Stack configuration for Apache.
 cd $out
-# These files are replaced by customizations.
-mv conf/httpd.conf conf/httpd.conf.built
-mv conf/mime.types conf/mime.types.built
-mv conf/magic conf/magic.built
 
 # Our own httpd.conf is fundamental to the whole idea.
+mv conf/httpd.conf conf/httpd.conf.built
 cp $httpdConf ./conf/httpd.conf
-chmod u+w ./conf/httpd.conf
-# Use '|' because '$out' contains forward slashes. Need to double
-# escape once for bash, then once for perl. I originally tried to use
-# the 'builtins' 'readFile' and 'toFile' to process the template
-# httpd.conf in the Nix expression, but was having trouble getting it
-# to work. This is a bit less elegant, but expedient.
-perl -p -i.bak -e "s|\\$\\{out\\}|$out|g" ./conf/httpd.conf
-# After this setup, the master 'httpd.conf' should only be changed
-# with a new apache installation or update. Individual sites
-# (VirtualHosts) may be changed by manipulating files in the
-# 'runtime-conf' which is included from the master file.
-echo "Include $DATA_DIR/runtime-conf/*.httpd.conf" >> ./conf/httpd.conf
-chmod u-w ./conf/httpd.conf
+# Also copy over our conveyor control scripts.
+cp $bin/* ./bin
 
 # These two files fix issues we had at one time. At
 # this point, I don't remember exactly what they are.
@@ -44,29 +34,35 @@ chmod u-w ./conf/httpd.conf
 # However, it would be good to either a) document why
 # they're necessary or b) see about using the default
 # again.
+mv conf/mime.types conf/mime.types.built
 cp $httpdMime ./conf/mime.types
+mv conf/magic conf/magic.built
 cp $httpdMagic ./conf/magic
 
 # Setup the Conveyor runtime directories. Note '$home', not $HOME. Rather
 # than use 'impure' vars (imported from the environment used to launch
 # nix), the 'home' is an argument provided by the nix installation
 # expression.
-DATA_DIR=$home/.conveyor/data/conveyor-apache
+DATA_DIR=$home/.conveyor/data/dogfoodsoftware.com/conveyor/distro/pkgs/servers/http/conveyor-apache/
 # Note, the data dir may already exist; for instance, the package was
 # deleted then re-installed from nix.
 mkdir -p $DATA_DIR
 # These three directories are refenced by the httpd.conf file.
-mkdir -p $DATA_DIR/runtime-conf
+mkdir -p $DATA_DIR/conf-inc
 mkdir -p $DATA_DIR/misc # What's this for?
 mkdir -p $DATA_DIR/ssl
 # 'htdocs' to be removed; replaced by 'conveyor-host' package.
 mkdir -p $DATA_DIR/htdocs
 mv ./htdocs/* $DATA_DIR/htdocs
 rmdir ./htdocs
-ln -s $DATA_DIR/htdocs htdocs
+if [ ! -f $DATA_DIR/htdocs ]; then
+    ln -s $DATA_DIR/htdocs htdocs
+fi
 
 # We want logs to be both standard and durable, so we create
 # our own and then symlink the directory in the nix installation.
-mkdir $DATA_DIR/logs
+mkdir -p $DATA_DIR/logs
 rmdir ./logs
-ln -s $DATA_DIR/logs logs
+if [ ! -f $DATA_DIR/logs ]; then
+    ln -s $DATA_DIR/logs logs
+fi
